requirements.txt
Flask==2.0.1
Gunicorn==20.1.0
Jinja2==3.0.1
python-dotenv==0.20.0
